Fixed NPE in ModuleMaker.compile()

+++ core/kernel/source/jetbrains/mps/make/ModuleMaker.java
@@ -163,10 +163,12 @@ public class ModuleMaker {
         fqName = fqName.substring(0, fqName.length() - toCopy.getFile().getName().length());
         String path = fqName.replace('/', File.separatorChar) + toCopy.getFile().getName();
 
-        FileUtil.copyFile(
-          new File(toCopy.getFile().getAbsolutePath()),
-          new File(module.getClassesGen().child(path).getAbsolutePath())
-        );
+        if (new File(toCopy.getFile().getAbsolutePath()).exists()) {
+          FileUtil.copyFile(
+            new File(toCopy.getFile().getAbsolutePath()),
+            new File(module.getClassesGen().child(path).getAbsolutePath())
+          );
+        }
       }
     }
